pyproject.toml
[project]
name = "example-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[extensions]>=1.25.0",
    "python-dotenv>=1.2.1",
]
